Make metrics cards for service queues into extensions
