--- packages/esm-appointments-app/src/metrics/appointments-metrics.scss
@@ -1,14 +1,0 @@
-@use '@carbon/colors';
-@use '@carbon/layout';
-
-.cardContainer {
-  background-color: colors.$white;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(18.75rem, 1fr));
-  gap: layout.$spacing-05;
-  margin: layout.$spacing-05;
-}
-
-.errorContainer {
-  margin: layout.$spacing-05;
-}
