+++ packages/esm-appointments-app/src/metrics/metrics-container.scss
@@ -1,0 +1,22 @@
+@use '@carbon/layout';
+@use '@openmrs/esm-styleguide/src/vars' as *;
+
+.cardContainer {
+  background-color: $ui-02;
+  display: flex;
+  padding: 0 layout.$spacing-05 layout.$spacing-07 layout.$spacing-05;
+  flex-flow: row wrap;
+  margin-top: -(layout.$spacing-03);
+  gap: layout.$spacing-05;
+}
+
+.cardContainer > * {
+  flex: 1 0 0%;
+  min-width: 18.75rem;
+}
+
+@media (max-width: calc(layout.$spacing-05 * 2 + 18.75rem)) {
+  .cardContainer > *:has(:global(.cardWithChildren)) {
+    order: 999;
+  }
+}
