+++ packages/esm-service-queues-app/src/home.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PatientQueueHeader from './patient-queue-header/patient-queue-header.component';
-import ClinicMetrics from './patient-queue-metrics/clinic-metrics.component';
+import ClinicMetrics from './metrics/metrics-container.component';
 import DefaultQueueTable from './queue-table/default-queue-table.component';
 
 const Home: React.FC = () => {
