+++ packages/esm-service-queues-app/src/metrics/metrics-cards/average-wait-time.extension.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import MetricsCard from './metrics-card.component';
+import { useAverageWaitTime } from '../metrics.resource';
+import { useSelectedService } from '../../helpers/helpers';
+
+export default function AverageWaitTimeExtension() {
+  const { t } = useTranslation();
+  const currentService = useSelectedService();
+  const { waitTime } = useAverageWaitTime(currentService?.serviceUuid, '');
+
+  return (
+    <MetricsCard
+      label={t('minutes', 'Minutes')}
+      headerLabel={t('averageWaitTime', 'Average wait time today')}
+      service="waitTime"
+      value={waitTime ? waitTime.averageWaitTime : '--'}
+    />
+  );
+}
