--- packages/esm-service-queues-app/src/patient-queue-metrics/clinic-metrics.scss
@@ -1,11 +1,0 @@
-@use '@carbon/layout';
-@use '@openmrs/esm-styleguide/src/vars' as *;
-
-.cardContainer {
-  background-color: $ui-02;
-  display: flex;
-  justify-content: space-between;
-  padding: 0 layout.$spacing-05 layout.$spacing-07 layout.$spacing-03;
-  flex-flow: row wrap;
-  margin-top: -(layout.$spacing-03);
-}
